Add shared UUID generator utility
